+++ projects/core/src/cms/cms.module.ts
@@ -1,11 +1,12 @@
 import { NgModule } from '@angular/core';
 
-import { CmsService } from './facade/index';
+import { CmsService } from './facade';
 import { CmsStoreModule } from './store/cms-store.module';
-import { CmsOccModule } from './occ/cms-occ.module';
+import { CmsOccModule } from './occ';
+import { ComponentMapperService, DefaultPageService } from './services';
 
 @NgModule({
   imports: [CmsOccModule, CmsStoreModule],
-  providers: [CmsService]
+  providers: [CmsService, DefaultPageService, ComponentMapperService]
 })
 export class CmsModule {}
