+++ projects/core/src/cms/facade/cms.service.spec.ts
@@ -11,7 +11,7 @@ import { of } from 'rxjs';
 import { Page } from '../model/page.model';
 import { ContentSlotData } from '../model/content-slot.model';
 
-import { DefaultPageService } from '../occ/default-page.service';
+import { DefaultPageService } from '../services';
 import { CmsConfig } from '../config/cms-config';
 import { PageType } from '../../occ/occ-models/occ.models';
