+++ projects/core/src/cms/index.ts
@@ -2,13 +2,13 @@ export * from './config/cms-config';
 export * from './model/page.model';
 export * from './model/node-item.model';
 export * from './model/content-slot.model';
-export * from './occ/index';
-export * from './occ/cms-occ.module';
+export * from './occ';
+export * from './services';
 
 export * from './store/cms-state';
-export * from './store/actions/index';
-export * from './store/selectors/index';
+export * from './store/actions';
+export * from './store/selectors';
 
-export * from './facade/index';
+export * from './facade';
 export * from './cms.module';
 export { defaultCmsModuleConfig } from './config/default-cms-config';
