+++ projects/core/src/cms/occ/cms-occ.module.ts
@@ -1,12 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
-import { OccCmsService } from './occ-cms.service';
-import { ComponentMapperService } from './component-mapper.service';
-import { DefaultPageService } from './default-page.service';
+
+import { OccCmsService } from './cms.service';
 
 @NgModule({
   imports: [CommonModule, HttpClientModule],
-  providers: [OccCmsService, ComponentMapperService, DefaultPageService]
+  providers: [OccCmsService]
 })
 export class CmsOccModule {}
