+++ projects/core/src/cms/occ/cms.service.ts
@@ -4,8 +4,8 @@ import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
 import { catchError } from 'rxjs/operators';
 import { IdList } from './../model/idList.model';
 import { CmsConfig } from '../config/cms-config';
-import { PageContext } from '../../routing/index';
-import { CMSPage, PageType, CmsComponent } from '../../occ/occ-models/index';
+import { PageContext } from '../../routing';
+import { CMSPage, PageType, CmsComponent } from '../../occ';
 
 @Injectable()
 export class OccCmsService {
