+++ projects/core/src/cms/occ/index.ts
@@ -1,3 +1,2 @@
-export * from './occ-cms.service';
-export * from './component-mapper.service';
-export * from './default-page.service';
+export * from './cms.service';
+export * from './cms-occ.module';
