+++ projects/core/src/cms/services/default-page.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed, inject } from '@angular/core/testing';
 import { DefaultPageService } from './default-page.service';
 import { CmsConfig } from '../config/cms-config';
-import { PageType } from '../../occ/occ-models/occ.models';
+import { PageType } from '../../occ';
 
 const MockCmsModuleConfig: CmsConfig = {
   defaultPageIdForType: {
