+++ projects/core/src/cms/services/index.ts
@@ -1,0 +1,2 @@
+export * from './component-mapper.service';
+export * from './default-page.service';
