+++ projects/core/src/cms/store/effects/navigation-entry-item.effect.ts
@@ -5,9 +5,9 @@ import { Observable, of } from 'rxjs';
 import { map, catchError, filter, mergeMap, take } from 'rxjs/operators';
 
 import * as navigationItemActions from '../actions/navigation-entry-item.action';
-import { OccCmsService } from '../../occ/occ-cms.service';
+import { OccCmsService } from '../../occ';
 import { IdList } from '../../model/idList.model';
-import { RoutingService } from '../../../routing/index';
+import { RoutingService } from '../../../routing';
 
 @Injectable()
 export class NavigationEntryItemEffects {
