+++ projects/core/src/cms/store/effects/page.effect.ts
@@ -13,14 +13,14 @@ import {
 
 import * as pageActions from '../actions/page.action';
 import * as componentActions from '../actions/component.action';
-import { OccCmsService } from '../../occ/occ-cms.service';
-import { DefaultPageService } from '../../occ/default-page.service';
+import { OccCmsService } from '../../occ';
+import { DefaultPageService } from '../../services';
 
 import { Page } from '../../model/page.model';
 import { ContentSlotData } from '../../model/content-slot.model';
 
-import { RoutingService, PageContext } from '../../../routing/index';
-import { PageType, CMSPage } from '../../../occ/occ-models/index';
+import { RoutingService, PageContext } from '../../../routing';
+import { PageType, CMSPage } from '../../../occ';
 
 @Injectable()
 export class PageEffects {
